Reorganize go code
