+++ configurator/configurator_test.go
