+++ configurator/testFixtures/honeypot_config_malformed.json
