+++ monitor/regex.txt
