+++ monitor/test.sh
